+++ R/wrangling_issues.R
@@ -16,50 +16,117 @@
 #' @export
 #'
 #' @examples
-#'
 #' # Empty issue
 #' issue1 <- new_issue()
 #'
 #' # Custom issue
-#' issue2 <- new_issue(
+#' issue1 <- new_issue(
 #'     title = "Nouvelle issue",
 #'     body = "Un nouveau bug pour la fonction...",
 #'     number = 47,
 #'     created_at = Sys.Date()
 #' )
 #'
-#' issue3 <- new_issue(issue = issue2)
-#'
-new_issue <- function(
+#' issue2 <- new_issue(x = issue1)
+new_issue <- function(x = NULL, ...) {
+    UseMethod("new_issue", x)
+}
+
+#' @rdname new_issue
+#' @exportS3Method new_issue IssueTB
+#' @method new_issue IssueTB
+#' @export
+new_issue.IssueTB <- function(x, ...) {
+    return(x)
+}
+
+#' @rdname new_issue
+#' @exportS3Method new_issue data.frame
+#' @method new_issue data.frame
+#' @export
+new_issue.data.frame <- function(x, ...) {
+    issue <- do.call(args = x, what = new_issue)
+    return(issue)
+}
+
+#' @rdname new_issue
+#' @exportS3Method new_issue list
+#' @method new_issue list
+#' @export
+new_issue.list <- function(x, ...) {
+    issue <- do.call(args = x, what = new_issue)
+    return(issue)
+}
+
+#' @rdname new_issue
+#' @exportS3Method new_issue IssuesTB
+#' @method new_issue IssuesTB
+#' @export
+new_issue.IssuesTB <- function(x, ...) {
+    if (nrow(x) != 1L) {
+        stop("There are several issues in the object `x`.")
+    }
+    return(NextMethod())
+}
+
+#' @rdname new_issue
+#' @exportS3Method new_issue default
+#' @method new_issue default
+#' @export
+new_issue.default <- function(
+    x,
     title,
     body,
     number,
     state = c("open", "closed"),
     created_at = Sys.Date(),
     labels = NULL,
-    milestone = NULL,
-    issue = list(),
-    repo = NULL,
-    owner = NULL,
+    milestone = NA_character_,
+    repo = NA_character_,
+    owner = NA_character_,
+    url = NA_character_,
+    html_url = NA_character_,
+    comments = NULL,
+    creator = NA_character_,
+    assignee = NA_character_,
     ...
 ) {
     state <- match.arg(state)
 
-    if (!(missing(title) || missing(body) || missing(number))) {
-        issue <- list(
-            title = title,
-            body = body,
-            number = as.integer(number),
-            created_at = format_timestamp(created_at),
-            labels = labels,
-            milestone = milestone,
-            repo = repo,
-            owner = owner
-        )
-    } else if (!missing(issue)) {
-        issue[["created_at"]] <- format_timestamp(issue[["created_at"]])
-        issue[["number"]] <- as.integer(issue[["number"]])
+    if (missing(title) && missing(body) && missing(number)) {
+        title <- character(0L)
+        body <- character(0L)
+        number <- integer(0L)
+        state <- character(0L)
+        created_at <- format_timestamp(as.Date(character(0L)))
+        labels <- list()
+        milestone <- character(0L)
+        repo <- character(0L)
+        owner <- character(0L)
+        url <- character(0L)
+        html_url <- character(0L)
+        comments <- list()
+        creator <- character(0L)
+        assignee <- character(0L)
     }
+
+    issue <- list(
+        number = as.integer(number),
+        title = title,
+        body = body,
+        state = state,
+        url = url,
+        html_url = html_url,
+        milestone = milestone,
+        created_at = format_timestamp(created_at),
+        creator = creator,
+        assignee = assignee,
+        owner = owner,
+        repo = repo,
+        labels = labels,
+        comments = comments
+    )
+
     class(issue) <- "IssueTB"
     return(issue)
 }
@@ -72,39 +139,38 @@ new_issue <- function(
 #' @export
 #'
 #' @examples
-#'
-#' # Empty issue
+#' # Empty list of issues
 #' issues1 <- new_issues()
 #'
-#' # Custom issue
-#' issues2 <- new_issues(
-#'     x = new_issue(
-#'         title = "Une autre issue",
-#'         body = "J'ai une question au sujet de...",
-#'         number = 2,
-#'         created_at = Sys.Date()
-#'     )
+#' # List of issues from issue
+#' issue1 <- new_issue(
+#'     title = "Une autre issue",
+#'     state = "open",
+#'     body = "J'ai une question au sujet de...",
+#'     number = 2,
+#'     created_at = Sys.Date()
+#' )
+#' issues2 <- new_issues(x = issue1)
+#'
+#' # Custom issues
+#' issues3 <- new_issues(
+#'     title = "Une autre issue",
+#'     state = "open",
+#'     body = "J'ai une question au sujet de...",
+#'     number = 2,
+#'     created_at = Sys.Date()
 #' )
 #'
-#' issues3 <- new_issues(x = list(
-#'     new_issue(
-#'         title = "Nouvelle issue",
-#'         body = "Un nouveau bug pour la fonction...",
-#'         state = "open",
-#'         number = 1,
-#'         created_at = Sys.Date()
-#'     ),
-#'     new_issue(
-#'         title = "Une autre issue",
-#'         body = "J'ai une question au sujet de...",
-#'         state = "closed",
-#'         number = 2,
-#'         created_at = Sys.Date()
-#'     )
-#' ))
+#' issues4 <- new_issues(
+#'     title = c("Nouvelle issue", "Une autre issue"),
+#'     body = c("Un nouveau bug pour la fonction...",  "J'ai une question au sujet de..."),
+#'     state = c("open", "closed"),
+#'     number = 1:2,
+#'     created_at = Sys.Date()
+#' )
 #' @rdname new_issues
 #'
-new_issues <- function(x = list()) {
+new_issues <- function(x = NULL, ...) {
     UseMethod("new_issues", x)
 }
 
@@ -112,40 +178,115 @@ new_issues <- function(x = list()) {
 #' @exportS3Method new_issues IssueTB
 #' @method new_issues IssueTB
 #' @export
-new_issues.IssueTB <- function(x) {
-    issues <- list(x)
-    return(new_issues(x = issues))
+new_issues.IssueTB <- function(x, ...) {
+    issues <- do.call(args = x, what = new_issues)
+    return(issues)
 }
 
 #' @rdname new_issues
 #' @exportS3Method new_issues IssuesTB
 #' @method new_issues IssuesTB
 #' @export
-new_issues.IssuesTB <- function(x) {
+new_issues.IssuesTB <- function(x, ...) {
     return(x)
+}
+
+#' @rdname new_issues
+#' @exportS3Method new_issues data.frame
+#' @method new_issues data.frame
+#' @export
+new_issues.data.frame <- function(x, ...) {
+    issues <- do.call(args = x, what = new_issues)
+    return(issues)
+}
+
+#' @rdname new_issues
+#' @exportS3Method new_issues list
+#' @method new_issues list
+#' @export
+new_issues.list <- function(x, ...) {
+    issues <- do.call(args = x, what = new_issues)
+    return(issues)
 }
 
 #' @rdname new_issues
 #' @exportS3Method new_issues default
 #' @method new_issues default
 #' @export
-new_issues.default <- function(x = list()) {
-    class(x) <- "IssuesTB"
-    return(x)
-}
+new_issues.default <- function(
+    x,
+    title,
+    body,
+    number,
+    state,
+    created_at = Sys.Date(),
+    labels = list(),
+    milestone = NA_character_,
+    repo = NA_character_,
+    owner = NA_character_,
+    url = NA_character_,
+    html_url = NA_character_,
+    comments = list(),
+    creator = NA_character_,
+    assignee = NA_character_,
+    ...
+) {
+    if (missing(title) && missing(body) && missing(number) && missing(state)) {
+        title <- character(0L)
+        body <- character(0L)
+        number <- integer(0L)
+        state <- character(0L)
+        created_at <- format_timestamp(as.Date(character(0L)))
+        labels <- list()
+        milestone <- character(0L)
+        repo <- character(0L)
+        owner <- character(0L)
+        url <- character(0L)
+        html_url <- character(0L)
+        comments <- list()
+        creator <- character(0L)
+        assignee <- character(0L)
+    }
 
-#' @exportS3Method `[[` IssuesTB
-#' @method `[[` IssuesTB
-#' @export
-`[[.IssuesTB` <- function(x, ...) {
-    return(new_issue(issue = NextMethod()))
+    if (length(labels) == 0) {
+        labels <- rep(list(NULL), times = length(title))
+    }
+    if (length(comments) == 0) {
+        comments <- rep(list(NULL), times = length(title))
+    }
+
+    issues <- data.frame(
+        number = as.integer(number),
+        title = title,
+        body = body,
+        state = state,
+        url = url,
+        html_url = html_url,
+        milestone = milestone,
+        created_at = format_timestamp(created_at),
+        creator = creator,
+        assignee = assignee,
+        owner = owner,
+        repo = repo,
+        stringsAsFactors = FALSE
+    )
+    issues$labels <- labels
+    issues$comments <- comments
+
+    class(issues) <- c("IssuesTB", "data.frame")
+
+    return(issues)
 }
 
 #' @exportS3Method `[` IssuesTB
 #' @method `[` IssuesTB
 #' @export
-`[.IssuesTB` <- function(x, ...) {
-    return(new_issues(NextMethod()))
+`[.IssuesTB` <- function(x, ..., drop = TRUE) {
+    output <- new_issues(NextMethod(object = x, generic = "IssuesTB"))
+    if (drop == TRUE && nrow(output) == 1L) {
+        return(new_issue(output))
+    }
+    return(output)
 }
 
 #' @exportS3Method `[<-` IssuesTB
@@ -181,11 +322,15 @@ append <- function(x, values, after = length(x)) {
 #' @param values a \code{IssueTB} or a \code{IssuesTB} object.
 #' @method append IssuesTB
 #' @export
-append.IssuesTB <- function(x, values, after) {
+append.IssuesTB <- function(x, values, after = length(x)) {
+
+    if (after > nrow(x)) after <- nrow(x)
+    if (after < 0L) after <- 0L
+
     if (inherits(values, "IssuesTB")) {
-        return(new_issues(NextMethod()))
+        return(rbind(x[seq_len(after), , drop = FALSE], values, x[-seq_len(after), , drop = FALSE]))
     } else if (inherits(values, "IssueTB")) {
-        return(append(x, values = new_issues(values)))
+        return(append(x, values = new_issues(values), after = after))
     } else {
         stop(
             "This function requires a IssueTB or IssuesTB object ",
@@ -206,17 +351,5 @@ append.default <- function(x, values, after = length(x)) {
 #' @method unique IssuesTB
 #' @export
 unique.IssuesTB <- function(x, incomparables = FALSE, ...) {
-    saved_issues <- new_issues()
-    for (issue in x) {
-        is_in <- FALSE
-        for (already in saved_issues) {
-            if (identical(issue, already)) {
-                is_in <- TRUE
-            }
-        }
-        if (!is_in) {
-            saved_issues <- append(saved_issues, issue)
-        }
-    }
-    return(saved_issues)
+    return(x[!duplicated(x), ])
 }