+++ R/wrangling_dataset_issues.R
@@ -67,20 +67,26 @@
 #'
 #' milestones <- get_milestones(source = "online")
 #' print(milestones)
-#'
+#' }
 #'
 #' # From local
 #'
-#' # First update the local database
-#' update_database(verbose = TRUE)
-#'
-#' issues <- get_issues(source = "local",
-#'                      dataset_name = "open_issues.yaml",
-#'                      state = "open")
-#' labels <- get_labels(source = "local")
-#' milestones <- get_milestones(source = "local")
-#' }
-#'
+#' path <- system.file("data_issues", package = "IssueTrackeR")
+#' issues <- get_issues(
+#'     source = "local",
+#'     dataset_dir = path,
+#'     dataset_name = "list_issues.yaml"
+#' )
+#' milestones <- get_issues(
+#'     source = "local",
+#'     dataset_dir = path,
+#'     dataset_name = "list_milestones.yaml"
+#' )
+#' labels <- get_issues(
+#'     source = "local",
+#'     dataset_dir = path,
+#'     dataset_name = "list_labels.yaml"
+#' )
 get_issues <- function(
     source = c("local", "online"),
     dataset_dir = getOption("IssueTrackeR.dataset.dir"),
@@ -356,16 +362,26 @@ format_issues <- function(
 #' @export
 #'
 #' @examples
-#' \donttest{
-#' all_issues <- get_issues(source = "online", verbose = FALSE)
+#' path <- system.file("data_issues", package = "IssueTrackeR")
+#' issues <- get_issues(
+#'     source = "local",
+#'     dataset_dir = path,
+#'     dataset_name = "list_issues.yaml"
+#' )
+#' milestones <- get_issues(
+#'     source = "local",
+#'     dataset_dir = path,
+#'     dataset_name = "list_milestones.yaml"
+#' )
+#' labels <- get_issues(
+#'     source = "local",
+#'     dataset_dir = path,
+#'     dataset_name = "list_labels.yaml"
+#' )
+#'
 #' write_issues_to_dataset(all_issues)
-#'
-#' labels <- get_labels(source = "online")
 #' write_labels_to_dataset(labels)
-#'
-#' milestones <- get_milestones(source = "online")
 #' write_milestones_to_dataset(milestones)
-#' }
 #'
 #' @rdname write
 #'