+++ NEWS.md
@@ -8,6 +8,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+* Url link with Issues printed
+
 ## [1.1.1] - 2025-04-26